# Fixes #20716: Improve dynamic group computation speed