--- a/webapp/sources/rudder/rudder-core/src/main/scala/com/normation/rudder/domain/queries/CmdbQuery.scala
+++ b/webapp/sources/rudder/rudder-core/src/main/scala/com/normation/rudder/domain/queries/CmdbQuery.scala
@@ -51,6 +51,7 @@ import com.normation.rudder.domain.nodes.NodeGroupId
 import com.normation.rudder.domain.nodes.NodeInfo
 import com.normation.rudder.domain.nodes.NodeState
 import com.normation.rudder.domain.properties.NodeProperty
+import com.normation.rudder.domain.queries.OstypeComparator.osTypes
 import com.normation.rudder.services.queries._
 import com.normation.zio._
 import com.unboundid.ldap.sdk._
@@ -221,8 +222,72 @@ final case object NodeStateComparator extends NodeCriterionType {
     )
 }
 
+final case object NodeOstypeComparator extends NodeCriterionType {
+  override def comparators = Seq(Equals, NotEquals)
+  override protected def validateSubCase(v:String,comparator:CriterionComparator) = {
+    if(null == v || v.isEmpty) Left(Inconsistency("Empty string not allowed")) else Right(v)
+  }
 
-final case object NodeStringComparator extends NodeCriterionType {
+  override def matches(comparator: CriterionComparator, value: String): NodeInfoMatcher = {
+    comparator match {
+      case Equals => NodeInfoMatcher(s"Prop equals '${value}'", (node: NodeInfo) => node.osDetails.os.kernelName == value )
+      case _      => NodeInfoMatcher(s"Prop not equals '${value}'", (node: NodeInfo) => node.osDetails.os.kernelName != value )
+    }
+  }
+
+  override def toForm(value: String, func: String => Any, attrs: (String, String)*) : Elem =
+    SHtml.select(
+      (osTypes map (e => (e,e))).toSeq
+      , { if(osTypes.contains(value)) Full(value) else Empty}
+      , func
+      , attrs:_*
+    )
+}
+
+final case object NodeOsNameComparator extends NodeCriterionType {
+
+  import net.liftweb.http.S
+
+  val osNames = AixOS ::
+    BsdType.allKnownTypes.sortBy { _.name } :::
+    LinuxType.allKnownTypes.sortBy { _.name } :::
+    (SolarisOS :: Nil) :::
+    WindowsType.allKnownTypes
+
+
+  override def comparators = Seq(Equals, NotEquals)
+  override protected def validateSubCase(v:String,comparator:CriterionComparator) = {
+    if(null == v || v.isEmpty) Left(Inconsistency("Empty string not allowed")) else Right(v)
+  }
+
+  override def matches(comparator: CriterionComparator, value: String): NodeInfoMatcher = {
+    comparator match {
+      case Equals => NodeInfoMatcher(s"Prop equals '${value}'", (node: NodeInfo) => node.osDetails.os.name == value )
+      case _      => NodeInfoMatcher(s"Prop not equals '${value}'", (node: NodeInfo) => node.osDetails.os.name != value )
+    }
+  }
+
+  private[this] def distribName(x: OsType): String = {
+    x match {
+      //add linux: for linux
+      case _: LinuxType   => "Linux - " + S.?("os.name."+x.name)
+      case _: BsdType     => "BSD - " + S.?("os.name."+x.name)
+      //nothing special for windows, Aix and Solaris
+      case _              => S.?("os.name."+x.name)
+    }
+  }
+
+  override def toForm(value: String, func: String => Any, attrs: (String, String)*) : Elem =
+    SHtml.select(
+      osNames.map(e => (e.name,distribName(e))).toSeq,
+      {osNames.find(x => x.name == value).map( _.name)},
+      func,
+      attrs:_*
+    )
+}
+
+
+final case class NodeStringComparator(access: NodeInfo => String) extends NodeCriterionType {
   override val comparators = BaseComparators.comparators
 
   override protected def validateSubCase(v: String, comparator: CriterionComparator) = {
@@ -235,13 +300,14 @@ final case object NodeStringComparator extends NodeCriterionType {
   }
 
   override def matches(comparator: CriterionComparator, value: String): NodeInfoMatcher = {
+
     comparator match {
-      case Equals    => NodeInfoMatcher(s"Prop equals '${value}'", (node: NodeInfo) => node.hostname == value )
-      case NotEquals => NodeInfoMatcher(s"Prop not equals '${value}'", (node: NodeInfo) => node.hostname != value )
-      case Regex     => NodeInfoMatcher(s"Prop matches regex '${value}'", (node: NodeInfo) => node.hostname.matches(value) )
-      case NotRegex  => NodeInfoMatcher(s"Prop matches not regex '${value}'", (node: NodeInfo) => !node.hostname.matches(value) )
-      case Exists    => NodeInfoMatcher(s"Prop exists", (node: NodeInfo) => node.hostname.nonEmpty )
-      case NotExists => NodeInfoMatcher(s"Prop doesn't exists", (node: NodeInfo) => node.hostname.isEmpty )
+      case Equals    => NodeInfoMatcher(s"Prop equals '${value}'", (node: NodeInfo) => access(node) == value )
+      case NotEquals => NodeInfoMatcher(s"Prop not equals '${value}'", (node: NodeInfo) => access(node) != value )
+      case Regex     => NodeInfoMatcher(s"Prop matches regex '${value}'", (node: NodeInfo) => access(node).matches(value) )
+      case NotRegex  => NodeInfoMatcher(s"Prop matches not regex '${value}'", (node: NodeInfo) => !access(node).matches(value) )
+      case Exists    => NodeInfoMatcher(s"Prop exists", (node: NodeInfo) => access(node).nonEmpty )
+      case NotExists => NodeInfoMatcher(s"Prop doesn't exists", (node: NodeInfo) => access(node).isEmpty )
 
     }
   }