https://issues.rudder.io/issues/20716

This is a WIP - the code is not finished, and may land in another branch

The goal here is to explore how to do dynamic group computation with minimal LDAP queries, which are really expensive
We have many data in the nodeinfocache, we can use them

As a first approach, I changed the ldap search on hostname to a nodeinfofilter (so that it's a *post* filter). The drawback is that when we only search on hostname, there is no filter to apply, so it fetches all nodes to do the search. Kind of kill the advantage of doing so, that's why i resolve to add a boolean that says: ok, don't compute anything, and take all from cache