@ncharles 2022-02-03
Impact on simple search on hostname is dramatic: from 500ms to nearly nothing when the search don't return any node (the display part is really expensive on 10 000 nodes, so i can only measure when nothing is returned)