@ncharles 2022-02-03
The queryChecker need also to be corrected/modified to make use of NodeInfoFilter, using the getPendingNodeInfos of NodeInfoService