@ncharles 2022-02-04
I managed, on my test bench, with the code at this point, to have a *2 perfs increase, from 2 minutes to 1 minute
I've been really surprised by the total time as most of my 497 query went from about 100-300ms to 30 ms
The sum of all the  `com.normation.rudder.services.queries.AcceptedNodesLDAPQueryProcessor - LDAP result: 10001 entries obtained in 236ms for query` went from 71507ms to 8935ms (so the LDAP part is nearly 10 times faster)
The sum of all the postfilter `filter:rudderNode] Found 15 nodes when filtering for info service existence and properties (294 ms)` has understandably went up, from 7636ms to 22636ms (*3) 


Investigating showed:
* the time spent computing the group is now 66850 ms
* the time spend computing and UPDATING the group is 86500 ms

so 1/3 of the time is to check that the group changed on ldap 
but thee are nearly 33s unaccounted for (8935 + 22636 << 66850 )