@ncharles 2022-02-04
Ok, i strongly suspect that looking in the nodeinfoservicecached fo matching nodeinfo, to get to LDAPNodeInfo, to convert thme to QueryResult (same looking object) to put them in a Seq, to finally get the NodeId is probably node the finest move in processOnlyId
Same for Process
Both could simply return NodeInfo