@ncharles 2022-02-06
Now, dynamic group update is 50s with this change on my test bench, so a bit better than *2 perf improvement
52s spend in updating, 40s in computing (still 12 seconds for updated that don't update anythingr) <- note it's ore than 50 because it's multi-threaded

Edit: the logger part add a massive overhead, removing the debug from com.normation.rudder.services.queries.AcceptedNodesLDAPQueryProcessor show 35s for dynamic group update, about *4 perf improvement