@ncharles 2022-02-07
Tests now pass, the "or" part was not correctly handled
However perfs have plummeted to 1M15s
This need investigation